backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pillow==10.2.0
torch==2.1.2
torchvision==0.16.2
onnxruntime==1.16.3
numpy==1.26.3
aiofiles==23.2.1
redis==5.0.1
celery==5.3.4
prometheus-client==0.19.0
loguru==0.7.2
httpx==0.26.0
python-dotenv==1.0.0
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
